Reorganize comparators / string prep.
This allows moving schema creation / loading to strings / files rather than trying to build them via PHP classes / constants. This will let me move the standard schemas to resource files and lets custom schemas be more easily managed. Will follow up to move the existing built schemas.
