+++ src/FreeDSx/Ldap/Schema/Matching/Comparator/DistinguishedNameComparator.php
@@ -11,9 +11,11 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace FreeDSx\Ldap\Schema\Matching;
+namespace FreeDSx\Ldap\Schema\Matching\Comparator;
 
 use FreeDSx\Ldap\Entry\Dn;
+use FreeDSx\Ldap\Schema\Matching\MatchingRuleComparatorInterface;
+use FreeDSx\Ldap\Schema\Matching\SubstringAssertion;
 
 /**
  * DN equality comparator (distinguishedNameMatch): normalizes both sides before comparing.
