+++ src/FreeDSx/Ldap/Schema/Matching/Comparator/GeneralizedTimeComparator.php
@@ -11,10 +11,12 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace FreeDSx\Ldap\Schema\Matching;
+namespace FreeDSx\Ldap\Schema\Matching\Comparator;
 
 use DateTimeImmutable;
 use DateTimeZone;
+use FreeDSx\Ldap\Schema\Matching\MatchingRuleComparatorInterface;
+use FreeDSx\Ldap\Schema\Matching\SubstringAssertion;
 
 /**
  * Generalized time comparator (generalizedTimeMatch / generalizedTimeOrderingMatch).
