+++ src/FreeDSx/Ldap/Schema/Matching/Comparator/IntegerComparator.php
@@ -11,24 +11,21 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace FreeDSx\Ldap\Schema\Matching;
+namespace FreeDSx\Ldap\Schema\Matching\Comparator;
+
+use FreeDSx\Ldap\Schema\Matching\MatchingRuleComparatorInterface;
+use FreeDSx\Ldap\Schema\Matching\SubstringAssertion;
 
 /**
- * Bit-mask comparator for bitwise AND (bitAndMatch) and OR (bitOrMatch) matching rules.
+ * Integer comparator (integerMatch / integerOrderingMatch): compares string representations as integers.
  */
-final readonly class BitMaskComparator implements MatchingRuleComparatorInterface
+final class IntegerComparator implements MatchingRuleComparatorInterface
 {
-    public function __construct(private bool $requireAllBits) {}
-
     public function equals(
         string $a,
         string $b,
     ): bool {
-        $mask = (int) $b;
-
-        return $this->requireAllBits
-            ? ((int) $a & $mask) === $mask
-            : ((int) $a & $mask) !== 0;
+        return (int) $a === (int) $b;
     }
 
     public function compare(
