+++ src/FreeDSx/Ldap/Schema/Matching/Comparator/TelephoneNumberComparator.php
@@ -11,7 +11,10 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace FreeDSx\Ldap\Schema\Matching;
+namespace FreeDSx\Ldap\Schema\Matching\Comparator;
+
+use FreeDSx\Ldap\Schema\Matching\MatchingRuleComparatorInterface;
+use FreeDSx\Ldap\Schema\Matching\SubstringAssertion;
 
 /**
  * Telephone number comparator (telephoneNumberMatch): strips spaces and hyphens before comparing case-insensitively.
