+++ src/FreeDSx/Ldap/Server/Backend/Storage/FilterEvaluator.php
@@ -18,8 +18,8 @@ use FreeDSx\Ldap\Entry\Entry;
 use FreeDSx\Ldap\Exception\OperationException;
 use FreeDSx\Ldap\Operation\ResultCode;
 use FreeDSx\Ldap\Schema\Definition\SyntaxOid;
-use FreeDSx\Ldap\Schema\Matching\CaseIgnoreComparator;
-use FreeDSx\Ldap\Schema\Matching\IntegerComparator;
+use FreeDSx\Ldap\Schema\Matching\Comparator\CaseIgnoreComparator;
+use FreeDSx\Ldap\Schema\Matching\Comparator\IntegerComparator;
 use FreeDSx\Ldap\Schema\Matching\MatchingRuleComparatorInterface;
 use FreeDSx\Ldap\Schema\Matching\SubstringAssertion;
 use FreeDSx\Ldap\Schema\Schema;
